Add StateFunctor and StateDispatcher.
It is to allow filling (or precalculation) of State with data based
on Material and other Body components, like Shape.

The dispatching is done on the class State, and the purpose is to fill
State with correct data. Therefore correct State instance must already
exist, but it can be mostly uninitialised.