QMStateDiscrete::size wasn't updated correctly for 1D, 2D case.
The QMStateDiscrete::size isn't a Vector3r, because the wavefunction can
be a tensor product of several wavefunctions. So it's positional size
can have wildly varying number of dimensions.

However in some other places it is assumed that it's a Vector3r and
those must be fixed. Simplest it is to find them by looking through examples
and initialising
  halfSize  = [size_1d,size_1d]
instead of
  halfSize  = [size_1d,size_1d,0.2]
for 2D case.