FIXME? think: QMStateDiscrete::start(), QMStateDiscrete::end()
Should those be relative or local? I have switched them to local, but
I'm not sure about that.

The reason for switching is that drawing in Gl1_QMGeometry is already
moved into local coords. And when start() and end() are not local, then
the wavefunction is drawn twice further away.

The reason for not switching is that I want two wavefunctions to
interact together, with their potential depending on their relative
placement.

However it seems that the second case can be managed properly by Law2_*
functor? I am still not sure. Needs more investigation.