Potential is implemented as Interaction; starts to work.
Potential is now working correctly using all the interaction classes, but
there are some problems:

1. duplicate St1_QMStateHarmonic::getValPos(..) ← to display potential using Gl1_QMGeometry
             Law2_QMIGeom_QMIPhysHarmonic::go   ← to actually calculate this potential

   both contain the same equation of potential!

   note:     display should be actually with Gl1_QMIPhys, or Gl1_QMIGeom

2. Global potential is calculated inside SchrodingerKosloffPropagator,
as a sum of all potentials, but this can't (?) work like that because the
wavefunction gets entangled (?). Hmm or maybe it can work like that, but
is simply not efficient - recalculating potential, when it's constant
all the time. Later, with electromagnetic waves, the potential
(interaction) will need to evolve according to Maxwell eqs. Hmm Law2_*
will have to take care of that. Interesting, because the potential will
evolve on its own accoring to the Maxwell eqs, and will keep doing so
even in the absence of any interacting particles. Like kind of Global
law for interactions, just like SchrodingerKosloffPropagator is a
propagation law for fermions.

3. I must make somwhere the global entangled wavefunction, and let
SchrodingerKosloffPropagator work only on this one.

4. Then displayed on screen will be only the marginal distributions of
each particle, from the global entangled wavefunction