calcMarginalDistribution(..) can now normalize (default) the calculated distribution.
Also I am pondering on possible ways to actually calculate this marginal
distribution. Since the wavefunction is complex it is not immediately
obvious to me on how to do that. Currently it just works the same as with
Real numbers. But perhaps I'll see more when I'll be able to see
different projections of the same wavefunction (after Gl1_* will be able draw multiple).